Split classes in flashglobalization.* to other files
