+++ src/allclasses.cpp
@@ -70,8 +70,11 @@
 #include "scripting/flash/utils/IntervalManager.h"
 #include "scripting/flash/utils/IntervalRunner.h"
 #include "scripting/flash/geom/flashgeom.h"
-#include "scripting/flash/globalization/flashglobalization.h"
+#include "scripting/flash/globalization/datetimeformatter.h"
+#include "scripting/flash/globalization/datetimestyle.h"
+#include "scripting/flash/globalization/lastoperationstatus.h"
 #include "scripting/flash/globalization/currencyformatter.h"
+#include "scripting/flash/globalization/numberformatter.h"
 #include "scripting/flash/globalization/stringtools.h"
 #include "scripting/flash/external/ExternalInterface.h"
 #include "scripting/flash/media/flashmedia.h"
