+++ src/scripting/flash/globalization/datetimestyle.cpp
@@ -1,0 +1,36 @@
+/**************************************************************************
+    Lightspark, a free flash player implementation
+
+    Copyright (C) 2009-2013  Alessandro Pignotti (a.pignotti@sssup.it)
+
+    This program is free software: you can redistribute it and/or modify
+    it under the terms of the GNU Lesser General Public License as published by
+    the Free Software Foundation, either version 3 of the License, or
+    (at your option) any later version.
+
+    This program is distributed in the hope that it will be useful,
+    but WITHOUT ANY WARRANTY; without even the implied warranty of
+    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+    GNU Lesser General Public License for more details.
+
+    You should have received a copy of the GNU Lesser General Public License
+    along with this program.  If not, see <http://www.gnu.org/licenses/>.
+**************************************************************************/
+
+#include "scripting/flash/globalization/datetimestyle.h"
+#include "scripting/class.h"
+#include "scripting/argconv.h"
+#include "scripting/toplevel/Date.h"
+#include <iomanip>
+
+using namespace lightspark;
+
+void DateTimeStyle::sinit(Class_base* c)
+{
+	CLASS_SETUP_NO_CONSTRUCTOR(c, ASObject, CLASS_SEALED|CLASS_FINAL);
+	c->setVariableAtomByQName("CUSTOM",nsNameAndKind(),asAtomHandler::fromString(c->getSystemState(),"custom"),CONSTANT_TRAIT);
+	c->setVariableAtomByQName("LONG",nsNameAndKind(),asAtomHandler::fromString(c->getSystemState(),"long"),CONSTANT_TRAIT);
+	c->setVariableAtomByQName("MEDIUM",nsNameAndKind(),asAtomHandler::fromString(c->getSystemState(),"medium"),CONSTANT_TRAIT);
+	c->setVariableAtomByQName("NONE",nsNameAndKind(),asAtomHandler::fromString(c->getSystemState(),"none"),CONSTANT_TRAIT);
+	c->setVariableAtomByQName("SHORT",nsNameAndKind(),asAtomHandler::fromString(c->getSystemState(),"short"),CONSTANT_TRAIT);
+}
